% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binding_scores.R
\name{binding_scores}
\alias{binding_scores}
\title{Calculate binding scores for each gene in a SeRP experiment}
\usage{
binding_scores(
  data,
  sample1,
  sample2,
  bin,
  window_size,
  skip_5prime = 0,
  skip_3prime = 0,
  conf.level = 0.95,
  bpparam = BiocParallel::bpparam()
)
}
\arguments{
\item{data}{A \code{serp_data} object.}

\item{sample1}{Name of the first sample (the numerator). If missing, the default sample1 of the data set
will be used.}

\item{sample2}{Name of the second sample (the denominator). If missing, the default sample2 of the data set
will be used.}

\item{bin}{Binning mode (\code{bynuc} or \code{byaa}). If missing, the default binning level of the data set
will be used.}

\item{window_size}{Neighborhood size for the confidence interval calculation in nucleotides. If missing, the default
window size of the data set will be used.}

\item{skip_5prime}{How many nucleotides to skip at the 5' end of the ORF. Useful if you know that the 5' end
contains artifacts.}

\item{skip_3prime}{How many nucleotides to skip at the 3' end of the ORF. useful if you know that the 3' end
contains artifacts.}

\item{conf.level}{Confidence level.}

\item{bpparam}{A \code{\link[BiocParallel]{BiocParallelParam-class}} object.}
}
\value{
A \link[tibble]{tibble} with the following columns: \describe{
     \item{gene}{The gene/ORF name.}
     \item{exp}{The experiment name.}
     \item{rep}{The replicate name.}
     \item{score_position}{Position within the gene for which the confidence interval is returned. Corresponds
         to the position where the highest value of the lower CI bound was observed. If \code{bin == 'byaa'}
         this is measured in codons, otherwise in nucleotides.}
     \item{lo_CI}{Lower confidence bound at the position \code{score_position}.}
     \item{hi_CI}{Upper confidence bound at the position \code{score_position}.}
     \item{mean}{Estimated enrichment at the position \code{score_position}.}
     \item{total_enrichment}{Total enrichment of the gene, calculated using all reads mapped anywhere within
         the ORF}
     \item{sample1_total_counts}{Total number of reads mapped to this ORF. Note that the actual column name
         is the value of \code{sample1}.}
     \item{sample2_total_counts}{Total number of reads mapped to this ORF. Note that the actual column name
         is the value of \code{sample2}.}
     \item{sample1_total_RPM}{\code{sample1_total_counts} normalized to the number of mapped reads. Note that
         the actual column name is the value of \code{sample1}.}
     \item{sample2_total_RPM}{\code{sample2_total_counts} normalized to the number of mapped reads. Note that
         the actual column name is the value of \code{sample2}.}
     \item{sample1_avg_read_density}{Average read density in sample 1, calculated as
         \eqn{\frac{\sum_{i=1}^n \textrm{sample1}_i}{n}}{sum(sample1)/lengh}. If \code{bin} is \code{byaa},
         density is in reads/nucleotide, otherwise in reads/codon. Note that the actual column
         name is the value of \code{sample1}.}
     \item{sample2_avg_read_density}{Average read density in sample 2, calculated as
         \eqn{\frac{\sum_{i=1}^n \textrm{sample2}_i}{n}}{sum(sample2)/lengh}. If \code{bin} is \code{byaa},
         density is in reads/nucleotide, otherwise in reads/codon. Note that the actual column
         name is the value of \code{sample2}.}
     \item{rank}{Ranking of the gene within the experiment and replicate. Genes are ranked by \code{lo_CI}
         in descending order.}
}
}
\description{
A binding score for a gene is defined as the highest value of the position-wise confidence interval for the
ratio \eqn{\frac{\textrm{sample1}}{\textrm{sample2}}}{sample1/sample2}. Enrichment confidence intervals are
calculated with \link{binom_ci_profile} separately for each experiment and replicate. Calculated values for
all replicates are averaged to create a new replicate \code{avg} which is returned along with values for the
individual replicates.
}
\seealso{
\code{\link{defaults}}
}
