% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_enrichment.R
\name{binom_ci}
\alias{binom_ci}
\title{Calculate enrichment confidence interval.}
\usage{
binom_ci(sample1, sample2, sample1_total, sample2_total,
  conf.level = 0.95)
}
\arguments{
\item{sample1}{Read counts in sample1.}

\item{sample2}{Read counts in sample2.}

\item{sample1_total}{Total number of reads in sample1 (for library size normalization).}

\item{sample2_total}{Total number of reads in sample2 (for library size normalization).}

\item{conf.level}{Confidence level.}
}
\value{
A \link[tibble]{tibble} containing the observed proportions and the lower
and upper bounds of the confidence interval.
}
\description{
Given total read counts in two samples, calculates library size-normalized confidence intervals
for the ratio \eqn{\frac{\textrm{sample1}}{\textrm{sample2}}}{sample1/sample2}.
}
\details{
Uses the Agresti-Coull approximation to calculate binomial confidence intervals of
\eqn{\frac{\textrm{sample1}}{\textrm{sample1} + \textrm{sample2}}}{sample1/(sample1 + sample2)}.
These values are then converted to CIs of \eqn{\frac{\textrm{sample1}}{\textrm{sample2}}}{sample1/sample2}
and normalized for library size.
}
