% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_enrichment.R
\name{binom_ci_profile}
\alias{binom_ci_profile}
\title{Calculate position-wise enrichment confidence intervals}
\usage{
binom_ci_profile(data, gene, sample1, sample2, exp, rep, bin, window_size,
  conf.level = 0.95)
}
\arguments{
\item{data}{A \code{serp_data} object.}

\item{gene}{Name of the gene/ORF.}

\item{sample1}{Sample name to use in the numerator.}

\item{sample2}{Sample name to use in the denominator.}

\item{exp}{Experiment name(s). If missing, all experiments will be used.}

\item{rep}{Replicate name(s). If missing, all replicates will be used.}

\item{bin}{Binning to use. If missing, will be automatically determined.}

\item{window_size}{Neighborhood size in nucleotides. Will be automatically converted to codons if binning
by codoons}

\item{conf.level}{Confidence level.}
}
\value{
A \link[tibble]{tibble} with the following columns: \describe{
     \item{winmid}{Center of the neighborhood for which the CI was calculated.}
     \item{sample1}{Raw counts of sample1 at position \code{winmid}. Note that the actual column name
             is the value of \code{sample1}.}
     \item{sample2}{Raw counts of sample2 at position \code{winmid}. Note that the actual column name
             is the value of \code{sample2}.}
     \item{win_sample1}{Total read counts of sample1 within the neighborhood centered at \code{winmid}.
             Note that the actual column name is the value of \code{sample1}.}
     \item{win_sample2}{Total read counts of sample2 within the neighborhood centered at \code{winmid}.
             Note that the actual column name is the value of \code{sample2}.}
     \item{ratio_mean}{Estimated ratio \eqn{\frac{sample1}{sample2}}.}
     \item{lo_CI}{Lower confidence bound.}
     \item{hi_CI}{Upper confidence bound.}
}
}
\description{
For a given gene, confidence intervals of enrichment (read count ratio) are calculated for each position.
}
\details{
At each position within the gene, read counts within a \code{window_size}-wide neighborhood are summed up
and used for CI calculation. A confidence interval for the ratio \eqn{\frac{\textrm{sample1}}{\textrm{sample2}}}{sample1/sample2}
is calculated using \link{binom_ci}.
}
