% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{downsample}
\alias{downsample}
\title{Downsample all samples to the same total counts}
\usage{
downsample(data)
}
\arguments{
\item{data}{A \code{serp_data} object}
}
\value{
A \code{serp_data} object
}
\description{
The minimum total count of every sample and sample type is used as downsampling target.
}
