% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binding_scores.R
\name{fit_background_model}
\alias{fit_background_model}
\title{Fit a background model to a selective ribosome profiling data set}
\usage{
fit_background_model(data, sample1, sample2, bin, tunnelcoords = 18:90,
  use_quantile = 0.5)
}
\arguments{
\item{data}{A \code{serp_data} object.}

\item{sample1}{Name of the first sample (the numerator). If missing, the default sample1 of the data set
will be used.}

\item{sample2}{Name of the second sample (the denominator). If missing, the default sample2 of the data set
will be used.}

\item{bin}{Binning mode (\code{bynuc} or \code{byaa}). If missing, the default binning level of the data set
will be used.}

\item{tunnelcoords}{Ribosome positions in nucleotites for which it is assumed that the nascent chain is
fully obscured by the exit tunnel.}

\item{quantile}{Only genes whith \code{sample2 >= quantile(sample2, use_quantile)} will be used for
parameter estimation. This reduces the effect of random drop-outs on the final model.}
}
\value{
A \code{serp_data} object.
}
\description{
Reads in \code{sample1} at ribosome positions for which the nascent chain is completely obscured
by the ribosome exit tunnel (typically the first 30 codons of a CDS) are assumed to represent background
binding. A background distribution is fitted to read counts within the tunnel.
}
\details{
For each gene, read counts within \code{tunnelcoords} are summed up (separately for \code{sample1} and
\code{sample2}). Counts in \code{sample1} conditioned on \code{sample1} + \code{sample2} are assumed
to follow a beta-binomial distribution. Parameter estimation is performed by maximum likelihood separately
for each sample and replicate.
}
\seealso{
\code{\link{get_background_model}}, \code{\link{test_binding}}, \code{\link{get_binding_positions}}
}
