% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binding_scores.R
\name{get_binding_positions}
\alias{get_binding_positions}
\title{Extract statistically significant binding regions}
\usage{
get_binding_positions(data, fdr = 0.01)
}
\arguments{
\item{data}{A \code{serp_data} object. \code{\link{test_binding}} must have been run on the data.}

\item{fdr}{False discovery rate.}
}
\value{
A \link[tibble]{tibble} with the following columns: \describe{
     \item{exp}{Experiment.}
     \item{rep}{Replicate.}
     \item{gene}{Gene.}
     \item{start}{Start of continuous binding region. If the binning mode for \code{\link{fit_background_model}}
          was \code{byaa}, this will be in codons, otherwise in nucleotides.}
     \item{end}{End of continuous binding region.}
     \item{width}{Width of continuous binding region.}
     \item{sample1}{Sum of counts of sample1 within the binding region. Note that the actual column name
             is the value of \code{sample1} given to \code{\link{fit_background_model}}.}
     \item{sample2}{Sum of counts of sample2 within the binding region. Note that the actual column name
             is the value of \code{sample1} given to \code{\link{fit_background_model}}.}
}
}
\description{
Calculates continuous binding regions at given FDR.
}
\seealso{
\code{\link{fit_background_model}}, \code{\link{test_binding}}, \code{\link{plot_binding_positions}}
}
