% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binding_scores.R
\name{get_binding_positions_by_threshold}
\alias{get_binding_positions_by_threshold}
\title{Extract binding regions based on a threshold}
\usage{
get_binding_positions_by_threshold(
  data,
  sample1,
  sample2,
  bin,
  window_size,
  skip_5prime = 0,
  skip_3prime = 0,
  conf.level = 0.95,
  bpparam = BiocParallel::bpparam(),
  threshold = 1
)
}
\arguments{
\item{data}{A \code{serp_data} object.}

\item{sample1}{Name of the first sample (the numerator). If missing, the default sample1 of the data set
will be used.}

\item{sample2}{Name of the second sample (the denominator). If missing, the default sample2 of the data set
will be used.}

\item{bin}{Binning mode (\code{bynuc} or \code{byaa}). If missing, the default binning level of the data set
will be used.}

\item{window_size}{Neighborhood size for the confidence interval calculation in nucleotides. If missing, the default
window size of the data set will be used.}

\item{skip_5prime}{How many nucleotides to skip at the 5' end of the ORF. Useful if you know that the 5' end
contains artifacts.}

\item{skip_3prime}{How many nucleotides to skip at the 3' end of the ORF. useful if you know that the 3' end
contains artifacts.}

\item{conf.level}{Confidence level.}

\item{bpparam}{A \code{\link[BiocParallel]{BiocParallelParam-class}} object.}

\item{threshold}{The enrichment threshold.}
}
\value{
A \link[tibble]{tibble} with the following columns: \describe{
     \item{exp}{Experiment.}
     \item{rep}{Replicate.}
     \item{gene}{Gene.}
     \item{start}{Start of continuous binding region. If the binning mode for \code{\link{fit_background_model}}
          was \code{byaa}, this will be in codons, otherwise in nucleotides.}
     \item{end}{End of continuous binding region.}
     \item{width}{Width of continuous binding region.}
     \item{sample1}{Sum of counts of sample1 within the binding region. Note that the actual column name
             is the value of \code{sample1}.}
     \item{sample2}{Sum of counts of sample2 within the binding region. Note that the actual column name
             is the value of \code{sample1}.}
     \item{binding_class}{Binding class of the binding region. 2 for high-confidence regions, 1 for low-confidence regions.}
}
}
\description{
Calculates binding regions given a threshold. For each gene, enrichment confidence intervals over
smoothed read counts are calculated by \code{\link{binom_ci_profile}}. If the lower confidence bound
is above the threshold at a position, this position is considered bound with high confidence. If the
threshold lies within the confidence interval for a position, this position is considered bound with
low confidence. Positions with the upper bound of the confidence interval below the threshold are
considered unbound.
}
\seealso{
\code{\link{get_binding_positions}}, \code{\link{plot_binding_positions}}
}
