% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene.R
\name{make_aligned_mats}
\alias{make_aligned_mats}
\title{Align data matrices}
\usage{
make_aligned_mats(data, pos, lengths, pwidth, filter = NULL,
  binwidth = 1, binmethod = c("sum", "mean"),
  bpparam = BiocParallel::bpparam())
}
\arguments{
\item{data}{Names list of data matrices.}

\item{pos}{Vector of positions.}

\item{lengths}{Named vector of gene/ORF lengths.}

\item{pwidth}{Width of the final matrices.}

\item{filter}{Genes to include. Defaults to all genes.}

\item{binwidth}{Bin width, if binning is desired.}

\item{binmethod}{Binning method. \code{sum}: Read counts within each bin are summed up; \code{mean}:
Read counts are averaged.}
}
\value{
Named list of matrices. Matrices are of width \eqn{2\cdot \textrm{pwidth} + 1}, with \code{pos} in
     column \eqn{\textrm{pwidth} + 1}.
}
\description{
Centers each gene at a given position within the gene.
}
