% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene.R
\name{make_average_profilefun}
\alias{make_average_profilefun}
\alias{make_average_profilefun.serp_data}
\alias{make_average_profilefun.serp_features}
\title{Generate a default function for metagene profiles that computes the position-wise arithmetic mean.}
\usage{
make_average_profilefun(data)

\method{make_average_profilefun}{serp_data}(data)

\method{make_average_profilefun}{serp_features}(data)
}
\arguments{
\item{data}{The data}
}
\value{
A function that can be passed as \code{profilefun} to \code{\link{metagene_profiles}}
}
\description{
Generate a default function for metagene profiles that computes the position-wise arithmetic mean.
}
\seealso{
\code{\link{metagene_profiles}}, \link{make_enrichment_profilefun},
         \code{\link{make_expression_normalizefun}}
}
