% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene.R
\name{make_enrichment_profilefun}
\alias{make_enrichment_profilefun}
\title{Generate the default function for metagene enrichment profiles}
\usage{
make_enrichment_profilefun(data, sample1, sample2)
}
\arguments{
\item{data}{A \code{serp_data} object}

\item{sample1}{Name of the first sample (the numerator). If missing, the default sample1 of the data set
will be used.}

\item{sample2}{Name of the second sample (the denominator). If missing, the default sample2 of the data set
will be used.}
}
\value{
A function that can be passed as \code{profilefun} to \code{\link{metagene_profiles}}
}
\description{
Position-wise meta-enrichment is calculated as\eqn{E_i = \frac{\sum_g \textrm{sample1}_{gi}}{\sum_g \textrm{sample2}_{gi}}}{E_i = sum(sample[,i]) / sum(sample[,i])}
where \eqn{i} is the position and \eqn{g} the gene.
}
\seealso{
\link{defaults}, \code{\link{metagene_profiles}}, \code{\link{make_average_profilefun}},
         \code{\link{make_expression_normalizefun}}
}
