% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene.R
\name{make_expression_normalizefun}
\alias{make_expression_normalizefun}
\title{Generate the default function expression-normalized for metagene profiles}
\usage{
make_expression_normalizefun(data, sample2)
}
\arguments{
\item{data}{A \code{serp_data} object}

\item{sample2}{Name of the sample which all other samples will be normalized against. If missing,
the default sample2 of the data set (usually the total translatome) will be used. If missing
and no default is set, each sample will be normalized to its own expression.}
}
\value{
A function that can be passed as \code{normalizefun} to \code{\link{metagene_profiles}}
}
\description{
Each gene is normalized to its total exression divided by its length.
}
\seealso{
\link{defaults}, \code{\link{metagene_profiles}}, \code{\link{make_average_profilefun}},
         \code{\link{make_enrichment_profilefun}}
}
