% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_enrichment.R
\name{make_label_fun}
\alias{make_label_fun}
\title{Create a mapping between experiment identifiers and names}
\usage{
make_label_fun(...)
}
\arguments{
\item{...}{Name-value pairs of strings. Argument names correspond to experiment IDs and values to
experiment names to be used e.g. in plot legends.}
}
\value{
A function that takes a character vector of experiment IDs and returns a character vector
     of experiment names. This function can be used e.g. as labeller in \code{\link[ggplot2]{facet_wrap}}
     and\code{\link[ggplot2]{facet_grid}}
}
\description{
Create a mapping between experiment identifiers and names
}
\examples{
     labelfun <- make_label_fun(id1='DnaK', id2='DnaK -tig')
     labelfun('id1')
}
