% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene.R
\name{metagene_profiles}
\alias{metagene_profiles}
\title{Calculate metagene profiles}
\usage{
metagene_profiles(data, profilefun, len, filter = NULL, bin = 1,
  binmethod = c("sum", "mean"), normalizefun = NULL,
  align = c("start", "stop"), nboot = 100, what = "byaa")
}
\arguments{
\item{data}{A \code{serp_data} object}

\item{profilefun}{Function that calculates a profile. Must accept named arguments for all sample types
present in the data set as well as \code{exp} (experiment name), \code{rep} (replicate name),
\code{bin} (bin width), \code{binmethod} (binning method), \code{align} (alignment). Must return
either a single numeric vector or a named list of numeric vectors.}

\item{len}{Length of the profile.}

\item{filter}{List of genes to include. Defaults to all genes.}

\item{bin}{Bin width.}

\item{binmethod}{How to bin the data. \code{sum}: Sums all read counts, \code{mean}: Averages read counts}

\item{normalizefun}{Function that normalizes the data pefore binning and profile calculation. Must accept
the same arguments as \code{profilefun}. Must return a named list of matrices.}

\item{align}{Alignment of the metagene profile, one of \code{start} (5' end) or \code{stop} (3' end).
Alternatively, a numeric vector of alignment positions for each gene in code{filter} can be given. In
this case, genes will be aligned to the given positions and the profile will be centered at 0, spanning
\code{len} positions in either direction.}

\item{nboot}{Number of bootstrap samples.}

\item{what}{Binning type to use, one of \code{byaa} or \code{bynuc}.}
}
\description{
Calculates a metagene profile from the full data set as well as bootstrapping samples (sampling genes).
Profiles are calculated separately for each experiment and replicate.
}
\details{
Count matrices are first filtered to contain only genes in \code{filter}. They are then passed to
\code{normalizefun} as named arguments, with names corresponding to sample type. If \code{align}
is one of \code{start} or \code{stop}, normalized count matrices are first aligned, then binned
and trimmed to \code{len} columns. If \code{align} is a vector of positions, count matrices are
centered using \code{\link{make_aligned_mats}} without binning for performance reasons, binning
is performed on the final profiles. The centered and trimmed count matrices are passed as named
arguments to \code{profilefun}, which calculates the final profiles. \code{profilefun} can return
either a single numeric vector, representing a single profile calculated from all count matrices,
or a named list of numeric vectors.
}
