% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binding_scores.R
\name{plot_background_model}
\alias{plot_background_model}
\title{Quality control plots for background model estimation}
\usage{
plot_background_model(data, type = c("density", "ratio"))
}
\arguments{
\item{data}{A \code{serp_data} object.}

\item{type}{Type of plot, one of \code{density} or \code{ratio}.}
}
\description{
Creates one plot per sample and replicate in a faceted display.
}
\details{
In \code{density} mode, plots a histogram of \eqn{\frac{\textrm{sample1}}{\textrm{sample1} + \textrm{sample2}}}{sample1/(sample1 + sample2)}
overlaid with the probability density function of a \link[=dbeta]{Beta distribution} using the
parameters estimated by \code{\link{fit_background_model}}.

In \code{ratio} mode, plots a scatterplot of \eqn{\frac{\textrm{sample1}}{\textrm{sample1} + \textrm{sample2}}}{sample1/(sample1 + sample2)}
over \code{sample2} overlaid with points randomly drawn from a beta-binomial distribution using
the parameters estimated by \code{\link{fit_background_model}}.
}
\seealso{
\code{\link{fit_background_model}}
}
