% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene.R
\name{plot_metagene_profiles}
\alias{plot_metagene_profiles}
\title{Plot metagene profiles}
\usage{
plot_metagene_profiles(df, exp = NULL, colaes = exp, align = "start",
  highlightregion = list(), highlightargs = list(),
  conf.level = 0.95, ci.alpha = 0.3, ylab = "enrichment")
}
\arguments{
\item{df}{A data frame created by \code{\link{metagene_profiles}}.}

\item{exp}{Experiments to plot. Defaults to all experiments.}

\item{colaes}{Variable to use for the color scale.}

\item{align}{Alignment to use in X axis label.}

\item{highlightargs}{Named list of additional parameters to pass to \code{\link[ggplot2]{annotate}}}

\item{conf.level}{Confidence level to plot.}

\item{ci.alpha}{Transparency level for the CI shading.}

\item{higlightregion}{Regions in the plot to be highlighted. Can be either a list containing two-element
numeric vectors or a two-column matrix. Rectangles spanning the entire height of the plot will be
drawn between the given x positions.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
Plots a metagene profile previously calculated with \code{\link{metagene_profiles}}. Lines show
the real profile using all data, shading indicates a bootstrapping-based confdience interval.
}
