% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_enrichment.R
\name{pois_ci}
\alias{pois_ci}
\title{Calculate read counts confidence interval.}
\usage{
pois_ci(sample, conf.level = 0.95)
}
\arguments{
\item{sample}{Read counts.}

\item{conf.level}{Confidence level.}
}
\value{
A \link[tibble]{tibble} containing the observed proportions and the lower
and upper bounds of the confidence interval.
}
\description{
Given total read counts in a sample, calculates library size-normalized confidence intervals.
}
\details{
Uses the exact Poisson confidence interval from \code{\link[base]{poisson.test}}.
}
