% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_enrichment.R
\name{pois_ci_profile}
\alias{pois_ci_profile}
\title{Calculate position-wise confidence intervals}
\usage{
pois_ci_profile(data, gene, samples, exp, rep, bin, window_size,
  conf.level = 0.95, ignore_genecol = FALSE)
}
\arguments{
\item{data}{A \code{serp_data} object.}

\item{gene}{Name of the gene/ORF.}

\item{samples}{Sample name. If missing, all samples will be used.}

\item{exp}{Experiment name(s). If missing, all experiments will be used.}

\item{rep}{Replicate name(s). If missing, all replicates will be used.}

\item{bin}{Binning to use. If missing, will be automatically determined.}

\item{window_size}{Neighborhood size in nucleotides. Will be automatically converted to codons if binning
by codons.}

\item{conf.level}{Confidence level.}

\item{ignore_genecol}{If \code{TRUE}, assumes that the given \code{gene} is the gene identifier contained
in the \code{gene} column of the reference table. If \code{FALSE}, assumes that the given \code{gene}
corresponds to the \code{genename_column} setting in the \link{defaults}.}
}
\value{
A \link[tibble]{tibble} with the following columns: \describe{
     \item{exp}{Experiment name.}
     \item{rep}{Replicate name.}
     \item{sample}{Sample name.}
     \item{winmid}{Center of the neighborhood for which the CI was calculated.}
     \item{counts}{Raw counts at position \code{winmid}.}
     \item{RPM}{RPM at position \code{winmid}.}
     \item{win_counts}{Total read counts within the neighborhood centered at \code{winmid}.}
     \item{win_RPM}{Total RPM within the neighborhood centered at \code{winmid}.}
     \item{counts_mean}{Estimated counts.}
     \item{RPM_mean}{Estimated RPM.}
     \item{lo_CI_counts}{Lower confidence bound for raw counts.}
     \item{hi_CI_counts}{Upper confidence bound for raw counts.}
     \item{lo_CI}{Lower confidence bound for RPM.}
     \item{hi_CI}{Upper confidence bound for RPM.}
}
}
\description{
For a given gene, confidence intervals of read counts are calculated for each position.
}
\details{
At each position within the gene, read counts within a \code{window_size}-wide neighborhood are summed up
and used for CI calculation. A confidence interval is calculated using \code{\link{pois_ci}}.
}
