% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{serp_data_accessors}
\alias{serp_data_accessors}
\alias{get_data.serp_data}
\alias{experiments}
\alias{samples}
\alias{get_reference.serp_data}
\alias{get_total_counts.serp_data}
\alias{get_defaults.serp_data}
\alias{is_normalized.serp_data}
\alias{is_downsampled}
\alias{excluded.serp_data}
\alias{get_background_model}
\alias{get_binding_pvalues}
\title{Accessors for \code{serp_data} objects}
\usage{
\method{get_data}{serp_data}(data)

experiments(data)

samples(data)

\method{get_reference}{serp_data}(data)

\method{get_total_counts}{serp_data}(data, as_df = FALSE)

\method{get_defaults}{serp_data}(data)

\method{is_normalized}{serp_data}(data)

is_downsampled(data)

\method{excluded}{serp_data}(data)

get_background_model(data)

get_binding_pvalues(data)
}
\arguments{
\item{data}{A \code{serp_data} object.}
}
\value{
\describe{
     \item{get_data}{Nested named list, with first level representing the experiment, second level
         the replicate, third level the sample type, and fourth level the binning. Count tables are
         sparse matrices with each row corresponding to an ORF.}
     \item{experiments}{Character vector containing names of experiments in the data set.}
     \item{samples}{Character vector containing names of all samples present in the data set.}
     \item{get_reference}{Reference data frame. Guaranteed to contain at least the following columns:
         \describe{
             \item{gene}{Gene/ORF name. Must match the names given in the read count tables.}
             \item{length}{ORF length in nucleotides.}
             \item{cds_length}{ORF length in codons.}}
         }
     \item{get_total_counts}{Nested named list containing toal read counts for each sample. If \code{as_df}
         is \code{TRUE}, returns a \code{\link[tibble]{tibble}} instead.}
     \item{get_defaults}{Named list of default parameters for this data set.}
     \item{get_background_model}{Background model estimated by \code{\link{fit_background_model}}.}
     \item{get_binding_pvalues}{Binding p-values calculated by \code{\link{test_binding}}.}
     \item{is_normalized}{A logical value indicating whether the data object contains raw or normalized
         read counts.}
     }
     \item{is_downsampled}{A logical value indicating whether the data object has been downsampled to
         the lowest total read counts by \code{\link{downsample}}.}
     \item{excluded}{Character vector of genes excluded from analyses.}
}
\description{
Accessors for \code{serp_data} objects
}
