% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{serp_features}
\alias{serp_features}
\title{Create a \code{serp_features} object}
\usage{
serp_features(..., ref, bin = c("bynuc", "byaa"), defaults = list())
}
\arguments{
\item{...}{Name-value pairs of matrices. The name of each argument denotes the feature type,
the value is a matrix with each row corresponding to a gene/ORF.}

\item{ref}{ref Reference data frame containing at least the following columns:
\describe{
    \item{gene}{Gene/ORF name. Must match the names given in the read count tables.}
    \item{length}{ORF length in nucleotides.}
}}

\item{bin}{Whether the features are per codon (\code{byaa}) or nucleotide (\code{bynuc}).}

\item{defaults}{Default parameters for the feature set.}
}
\value{
An object of class \code{serp_features}
}
\description{
This class holds position-wise features that are independent of experiment and replicate
}
\seealso{
\code{\link{serp_feature_accessors}}, \code{\link{defaults}}
}
