% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{set_defaults}
\alias{set_defaults}
\title{Set default parameters of a \code{serp_data} object}
\usage{
set_defaults(data, defaults)
}
\arguments{
\item{data}{A \code{serp_data} object.}

\item{defaults}{Named list of default parameters.}
}
\description{
Set default parameters of a \code{serp_data} object
}
