% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{get_genecounts}
\alias{get_genecounts}
\title{Calculate total read counts per gene}
\usage{
get_genecounts(data)
}
\arguments{
\item{data}{A \code{serp_data} object.}
}
\value{
A \link[tibble]{tibble} with columns \code{exp}, \code{rep}, \code{sample}, \code{counts},
     and \code{RPM}. If \code{\link[=defaults]{genename_column}} is not \code{"gene"}, the tibble will
     in additon contain \code{\link[=defaults]{genename_column}}.
}
\description{
Calculate total read counts per gene
}
