% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene.R
\name{metagene_profiles}
\alias{metagene_profiles}
\alias{metagene_profiles.serp_data}
\alias{metagene_profiles.serp_features}
\title{Calculate metagene profiles}
\usage{
metagene_profiles(data, ...)

\method{metagene_profiles}{serp_data}(
  data,
  profilefun,
  len,
  bin,
  filter = NULL,
  binwidth = 1,
  binmethod = c("sum", "mean"),
  normalizefun = NULL,
  align = "start",
  nboot = 100,
  bpparam = BiocParallel::bpparam()
)

\method{metagene_profiles}{serp_features}(
  data,
  profilefun,
  len,
  bin,
  filter = NULL,
  binwidth = 1,
  binmethod = c("sum", "mean"),
  normalizefun = NULL,
  align = "start",
  nboot = 100,
  bpparam = BiocParallel::bpparam()
)
}
\arguments{
\item{data}{The data.}

\item{profilefun}{Function that calculates a profile. Must accept named arguments for all sample types
present in the data set (matrices with genes in rows cropped to the profile region.
Missing values, e.g. if a gene ends within the profile region, are encoded as \code{NA}.)
as well as \code{exp} (experiment name), \code{rep} (replicate name),
\code{binwidth} (bin width), \code{binmethod} (binning method), \code{align} (alignment),
\code{lengths} (named vector of gene lengths). Must return either a single numeric vector or
a named list of numeric vectors.}

\item{len}{Length of the profile.}

\item{bin}{Bin level (\code{bynuc} or \code{byaa}). If missing, the default binning level of the data set
will be used}

\item{filter}{List of genes to include. Defaults to all genes.}

\item{binwidth}{Bin width.}

\item{binmethod}{How to bin the data. \code{sum}: Sums all read counts, \code{mean}: Averages read counts}

\item{normalizefun}{Function that normalizes the data pefore binning and profile calculation. Must accept
the same arguments as \code{profilefun}. Data matrices will contain data for entire genes.
The \code{lengths} argument is guaranteed to contain only genes present in the data matrices
in the same order.
Must return a named list of matrices.}

\item{align}{Alignment of the metagene profile, one of \code{start} (5' end) or \code{stop} (3' end).
Alternatively, a named numeric vector of alignment positions for each gene in \code{filter} can be given. In
this case, names must correspond to gene names and genes will be aligned to the given positions
and the profile will be centered at 0, spanning \code{len} positions in either direction.}

\item{nboot}{Number of bootstrap samples.}

\item{bpparam}{A \code{\link[BiocParallel]{BiocParallelParam-class}} object.}
}
\value{
A \link[tibble]{tibble} with the following columns: \describe{
     \item{id}{ID of the bootstrap sample}
     \item{pos}{Distance from the \code{align} position. If \code{bin == 'byaa'} this is measured in codons,
         otherwise in nucleotides.}
     \item{boot}{Logical, indicates whether this profile was generated from a bootstrap sample or from the
         complete data set.}
     \item{type}{Only if \code{profilefun} returns a list. Corresponds to the name of the list element
         containing the profile.}
     \item{summary}{The value returned by \code{profilefun}.}
}
}
\description{
Calculates a metagene profile from the full data set as well as bootstrapping samples (sampling genes).
Profiles are calculated separately for each experiment and replicate.
}
\details{
Data matrices are first filtered to contain the same set of genes (i.e. only the intersection of
genes contained in all matrices is retained). For \code{serp_data} objects, this filtering is performed
separately for each experiment and replicate. Data matrices are the filtered to contain only genes in
\code{filter}, if \code{filter is given}. Matrices are then passed to \code{normalizefun}
as named arguments, with names corresponding to sample type. If \code{align} is one of \code{start}
or \code{stop}, normalized data matrices are first aligned, then binned and trimmed to \code{len}
columns. If \code{align} is a vector of positions, data matrices are centered using
\code{\link{make_aligned_mats}} without binning for performance reasons, binning
is performed on the final profiles. The centered and trimmed data matrices are passed as named
arguments to \code{profilefun}, which calculates the final profiles. \code{profilefun} can return
either a single numeric vector, representing a single profile calculated from all data matrices,
or a named list of numeric vectors.
}
\seealso{
\link{defaults}, \code{\link{make_average_profilefun}},
         \code{\link{make_enrichment_profilefun}}, \code{\link{make_expression_normalizefun}}
}
