% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_enrichment.R
\name{plot.serp_data}
\alias{plot.serp_data}
\title{Plot an enrichment profile}
\usage{
\method{plot}{serp_data}(data, gene, sample1, sample2, exp, rep, bin,
  window_size, conf.level = 0.95, colaes = exp,
  highlightregion = list(), highlightargs = list())
}
\arguments{
\item{data}{A \code{serp_data} object. Must contain raw (unnormalized) read counts.}

\item{gene}{Name of the gene/ORF to plot.}

\item{sample1}{Name of the first sample (the numerator). If missing, the default sample1 of the data set
will be used.}

\item{sample2}{Name of the second sample (the denominator). If missing, the default sample2 of the data set
will be used.}

\item{exp}{Character vector of experiments to plot. If missing, all experiments are plotted.}

\item{rep}{Character vector of replicates to plot. If missing, all replicates will be plotted.}

\item{bin}{Bin level (\code{bynuc} or \code{byaa}). If missing, the default binning level of the data set
will be used.}

\item{window_size}{Neighborhood size for the confidence interval calculation in nucleotides. If missing, the default
window size of the data set will be used.}

\item{conf.level}{Confidence level.}

\item{colaes}{Variable to use for the color scale.}

\item{highlightargs}{Named list of additional parameters to pass to \code{\link[ggplot2]{annotate}}}

\item{higlightregion}{Regions in the plot to be highlighted. Can be either a list containing two-element
numeric vectors or a two-column matrix. Rectangles spanning the entire height of the plot will be
drawn between the given x positions.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
For each position, a binomial confidence interval calculated using \code{\link{binom_ci_profile}} is
plotted. Transparency reflects the total number of reads contributing to the confidence interval.
}
\seealso{
\code{\link{defaults}}
}
