% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binding_scores.R
\name{plot_binding_positions}
\alias{plot_binding_positions}
\title{Plot binding regions}
\usage{
plot_binding_positions(df, bgcolor = "lightgrey", ylabels = NA)
}
\arguments{
\item{df}{A data frame created by \code{\link{get_binding_positions}}.}

\item{bgcolor}{Background color to visualize unbound regions.}

\item{ylabels}{Whether to show Y axis labels (gene names). Defaults to suppressing Y axis labels if more than
10 genes are plotted.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
Plots chaperone binding regions in a heat map. Color intensity is proportional to the fraction of replicates
in which binding at the respective position was observed.
}
\seealso{
\code{\link{fit_background_model}}, \code{\link{test_binding}}, \code{\link{get_binding_positions}}
}
