% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{set_defaults}
\alias{set_defaults}
\alias{set_defaults.serp_data}
\alias{set_defaults.serp_features}
\title{Set default parameters}
\usage{
set_defaults(data, ...)

\method{set_defaults}{serp_data}(data, ...)

\method{set_defaults}{serp_features}(data, ...)
}
\arguments{
\item{data}{The data}

\item{...}{Name-value-pairs of parameters}
}
\description{
Set default parameters
}
